Refactor: - refactoring 'PlaceThumbnail.java' - rename 'PlaceThumbnailListResponse.java'
