+++ src/main/java/app/learningtrip/apiserver/course/domain/Course.java
@@ -1,7 +1,5 @@
 package app.learningtrip.apiserver.course.domain;
 
-import app.learningtrip.apiserver.dummy_temp.PlaceThumbnail;
-import app.learningtrip.apiserver.place.domain.Place;
 import app.learningtrip.apiserver.place.dto.response.PlaceResponse;
 import java.util.List;
 import lombok.Builder;
