+++ src/main/java/app/learningtrip/apiserver/dummy_temp/NearbyPlaceResponse.java
@@ -1,5 +1,6 @@
 package app.learningtrip.apiserver.dummy_temp;
 
+import app.learningtrip.apiserver.place.dto.response.PlaceThumbnail;
 import java.util.List;
 import lombok.Data;
