+++ src/main/java/app/learningtrip/apiserver/dummy_temp/SearchResultResponse.java
@@ -1,6 +1,7 @@
 package app.learningtrip.apiserver.dummy_temp;
 
 
+import app.learningtrip.apiserver.place.dto.response.PlaceThumbnail;
 import java.util.List;
 import lombok.Data;
