+++ src/main/java/app/learningtrip/apiserver/place/controller/PlaceController.java
@@ -1,14 +1,8 @@
 package app.learningtrip.apiserver.place.controller;
 
-import app.learningtrip.apiserver.dummy_temp.PlaceThumbnail;
-import app.learningtrip.apiserver.place.domain.Place;
-import app.learningtrip.apiserver.place.dto.response.PlaceDetailCultureResponse;
-import app.learningtrip.apiserver.place.dto.response.PlaceDetailTourResponse;
 import app.learningtrip.apiserver.place.dto.response.PlaceResponse;
-import app.learningtrip.apiserver.place.dto.response.PlaceThumbnailResponse;
+import app.learningtrip.apiserver.place.dto.response.PlaceThumbnailListResponse;
 import app.learningtrip.apiserver.place.service.PlaceService;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,8 +25,8 @@ public class PlaceController {
     }
 
     @GetMapping("/place/related/{place_id}")
-    public ResponseEntity<Optional<PlaceThumbnailResponse>> similarPlace(@PathVariable(name = "place_id") long place_id) {
-        Optional<PlaceThumbnailResponse> placeThumbnailResponse = Optional.ofNullable(placeService.similarPlaceDummy(place_id));
+    public ResponseEntity<Optional<PlaceThumbnailListResponse>> similarPlace(@PathVariable(name = "place_id") long place_id) {
+        Optional<PlaceThumbnailListResponse> placeThumbnailResponse = Optional.ofNullable(placeService.similarPlaceDummy(place_id));
         return ResponseEntity.ok().body(placeThumbnailResponse);
     }
 }
